refactor(Insurances): API呼び出し箇所をcomposablesとして抽出
Insurancesの責務は画面描画までとなるので、実際に処理を行う箇所については責務を分ける
またファイルを抽出することでテスタブルなコードにすることが目的

なお抽出に際して、Reactivity Transformにできるものは置き換えを行っている
